requirements.txt
streamlit==1.37.1
folium==0.17.0
streamlit-folium==0.20.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.8.0
